Update jquery.nav.js
The property I added (offsetFixer) is used in cases that you don't want to position the target section exactly on the top of page for some design reasons. In my case I have the Main menu with position fixed as result when you click on the link some of the content is hidden behind the menu and it is normal to do that. That's why I added this fixer to position the content -60px in my case. I think it's nice small future.